pull misnamed .cpp file into .h (#1270)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/1270

 Moves the entire contents of `CuptiActivity.cpp` into `CuptiActivity.h`.

The file `CuptiActivity.cpp` was not really a normal C++ source file. It was actually the inlined implementations for the generics declared in `CuptiActivity.h`. This meant that `CuptiActivityProfiler.cpp` had the line:
```
#include "CuptiActivity.cpp"
```
That is, at the least, confusing. The normal expectation for C++ code is that a `.h` is a header file that will be `#include`ed and a `.cpp` is a source file that will be compiled. I think we should stick to that.

Reviewed By: ryanzhang22, divyanshk

Differential Revision: D93888072

fbshipit-source-id: fd5994a502288c5a338ce12dc252814553575ccc
(cherry picked from commit 0f9bef9a7caf0c6529d43055cbcdb459f8224065)
